requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
attrs==21.2.0
black==21.5b2
click==8.0.1
Django==3.2.4
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.6.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pylint==2.8.3
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-factoryboy==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
wrapt==1.12.1
